Revert "move TaskFailedException to mrjob.sim"
This reverts commit d3a6deff12672e8613443009e0620869502039ab.
