+++ mrjob/sim.py
@@ -29,30 +29,9 @@ from mrjob.conf import combine_local_envs
 from mrjob.logs.counters import _format_counters
 from mrjob.parse import parse_mr_job_stderr
 from mrjob.runner import MRJobRunner
-from mrjob.step import StepFailedException
 from mrjob.util import unarchive
 
 log = logging.getLogger(__name__)
-
-
-class TaskFailedException(StepFailedException):
-    """Extension of :py:class:`~mrjob.step.StepFailedException` that blames
-    one particular task."""
-    _FIELDS = StepFailedException._FIELDS + ('task_type', 'input_path')
-
-    def __init__(
-            self, reason=None, step_num=None, num_steps=None, step_desc=None,
-            task_type=None, input_path=None):
-        super(TaskFailedException, self).__init__(
-            reason=reason, step_num=step_num,
-            num_steps=num_steps, step_desc=step_desc)
-
-        self.task_type = task_type
-        self.input_path = input_path
-
-    def __str__(self):
-        return (super(TaskFailedException, self).__str__() +
-            ' while reading from:\n  %s' % self.input_path)
 
 
 class SimMRJobRunner(MRJobRunner):
